Change the type of PW command number from uint to double [PLT/PLTDecoder.cs, PLT/Stroke.cs]

+++ GUI/PLT/Stroke.cs
@@ -6,13 +6,13 @@ namespace GUI.PLT {
         public Point2D Start { get; }
         public Point2D End { get; }
         public IColor StroceColor { get; }
-        public uint BrushWidth { get; }
+        public double BrushWidth { get; }
 
-        public Stroke(Point2D start, Point2D end, IColor color, uint brushWidth) {
+        public Stroke(Point2D start, Point2D end, IColor color, double brushWidth) {
             Start = start;
             End = end;
             StroceColor = color;
-            BrushWidth=brushWidth;
+            BrushWidth = brushWidth;
         }
     }
 }
